requirements.txt
dlib==19.24.0
face_recognition==1.3.0
imutils==0.5.4
numpy==1.15.1
numpy==1.21.5
opencv_contrib_python==4.7.0.72
opencv_python==4.7.0.72
Pillow==9.0.1
Pillow==9.5.0
scipy==1.7.3
